Add API v1 (#2595, #2425, #2331, #2066, #2062, #1762, #759, #394)
This is a major rework of the API with a lot of breaking changes. Compared the the old API v0, API v1:

- is generated using a Jekyll Generator (see https://jekyllrb.com/docs/plugins/generators/),
- is versioned using the api/v1 prefix (#2066) and a schema_version field (#2331). This will make it easier to implement non-backward-compatible changes in API,
- feels more "Restful",
- and expose most of the products / tags / categories data (#2062, #2595)

The API v0 is still generated to give time to users to migrate to API v1. It will be removed a year or more. The only change is that the 404 JSON redirect rule had to be reverted because of incompatibilities with API v1 own redirect rules (#2425).

The API v1 documentation is available at /docs/api/v1/, but the tool used for the documentation is now swagger-ui. It fixed the issues we had with stoplight webcomponent on CSP, code example generation and icon on Firefox ESR (#1762). But swagger-ui imposed the downgrade of OpenAPI version from 3.1.0 to 3.0.3.

Changes between the two versions has been documented in the API_CHANGELOG.md files.

Note that we thought of disabling API generation in development (using JEKYLL_ENV like the Jekyll Feed plugin - see https://github.com/jekyll/jekyll-feed/blob/master/lib/jekyll-feed/generator.rb#L145), but it was finally reverted. It does not work well with Netlify preview, and generate production URL (i.e. https://endoflife.date URLs) in development which makes it difficult to use.